pyproject.toml
[tool.poetry]
name = "python-stocks"
version = "0.1.0"
description = ""
authors = ["Lucas F. Silva <lucaslfsls@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.75.2"}
PyJWT = "^2.3.0"
cryptography = "^37.0.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = "^1.4.36"
asyncio = "^3.4.3"
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"
httpx = "^0.22.0"
mysql-connector-python = "^8.0.29"
asyncmy = "^0.2.5"
aiomysql = "^0.1.1"
yfinance = "^0.1.70"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]
pytest = {extras = ["asyncio"], version = "^7.1.2"}
pytest-asyncio = "^0.18.3"
jupyter = "^1.0.0"
Faker = "^13.7.0"
pandas = "^1.4.2"
openpyxl = "^3.0.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
